Refactored OauthUserAccountManagement.java to OAuthUserDetailsService
