+++ spring-boot/admin/src/main/java/eu/coatrack/admin/service/OAuthUserDetailsService.java
